@@ -78,7 +78,7 @@ namespace Examples.AdvancedProgramming.AsynchronousOperations
                     Console.WriteLine($"Addresses for {hostNames[i]}");
                     for (int k = 0; k < addresses.Length; k++)
                     {
-                        Console.WriteLine("{0}",addresses[k].ToString());
+                        Console.WriteLine($"{addresses[k].ToString()}");
                     }
                 }
             }
