@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 
 public class Example4
 {
@@ -9,8 +9,8 @@ public class Example4
         string fmt1 = "h \\h m \\m";
         string fmt2 = @"h \h m \m";
 
-        Console.WriteLine("{0} ({1}) -> {2}", date, fmt1, date.ToString(fmt1));
-        Console.WriteLine("{0} ({1}) -> {2}", date, fmt2, date.ToString(fmt2));
+        Console.WriteLine($"{date} ({fmt1}) -> {date.ToString(fmt1)}");
+        Console.WriteLine($"{date} ({fmt2}) -> {date.ToString(fmt2)}");
 
         // The example displays the following output:
         //       6/15/2009 1:45:30 PM (h \h m \m) -> 1 h 45 m
