@@ -1,4 +1,4 @@
-﻿// <Snippet3>
+// <Snippet3>
 using System;
 using System.Text;
 
@@ -19,7 +19,7 @@ public class Example
 
       byte[] bytes = cp1252r.GetBytes(str1);
       string str2 = cp1252r.GetString(bytes);
-      Console.WriteLine("Round-trip: {0}", str1.Equals(str2));
+      Console.WriteLine($"Round-trip: {str1.Equals(str2)}");
       if (! str1.Equals(str2)) {
          Console.WriteLine(str2);
          foreach (var ch in str2)
