@@ -1,4 +1,4 @@
-﻿// <Snippet7>
+// <Snippet7>
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -28,7 +28,7 @@ class Program
 
       // Decode the ASCII bytes.
       string str2 = enc.GetString(bytes);
-      Console.WriteLine("Round-trip: {0}", str1.Equals(str2));
+      Console.WriteLine($"Round-trip: {str1.Equals(str2)}");
       if (! str1.Equals(str2)) {
          Console.WriteLine(str2);
          foreach (var ch in str2)
