@@ -30,14 +30,9 @@ public class Example
       catch (EncoderFallbackException e) {
          Console.Write("Exception: ");
          if (e.IsUnknownSurrogate())
-            Console.WriteLine("Unable to encode surrogate pair 0x{0:X4} 0x{1:X3} at index {2}.",
-                              Convert.ToUInt16(e.CharUnknownHigh),
-                              Convert.ToUInt16(e.CharUnknownLow),
-                              e.Index);
+            Console.WriteLine($"Unable to encode surrogate pair 0x{Convert.ToUInt16(e.CharUnknownHigh):X4} 0x{Convert.ToUInt16(e.CharUnknownLow):X3} at index {e.Index}.");
          else
-            Console.WriteLine("Unable to encode 0x{0:X4} at index {1}.",
-                              Convert.ToUInt16(e.CharUnknown),
-                              e.Index);
+            Console.WriteLine($"Unable to encode 0x{Convert.ToUInt16(e.CharUnknown):X4} at index {e.Index}.");
          return;
       }
       Console.WriteLine();
@@ -45,7 +40,7 @@ public class Example
       // Decode the ASCII bytes.
       try {
          string str2 = enc.GetString(bytes);
-         Console.WriteLine("Round-trip: {0}", str1.Equals(str2));
+         Console.WriteLine($"Round-trip: {str1.Equals(str2)}");
          if (! str1.Equals(str2)) {
             Console.WriteLine(str2);
             foreach (var ch in str2)
