@@ -1,4 +1,4 @@
-﻿// <Snippet2>
+// <Snippet2>
 using System;
 using System.Text;
 
@@ -24,7 +24,7 @@ public class Example
 
       // Decode the ASCII bytes.
       string str2 = enc.GetString(bytes);
-      Console.WriteLine("Round-trip: {0}", str1.Equals(str2));
+      Console.WriteLine($"Round-trip: {str1.Equals(str2)}");
       if (! str1.Equals(str2)) {
          Console.WriteLine(str2);
          foreach (var ch in str2)
