@@ -9,8 +9,7 @@ public class Example
       DateTime date1 = new DateTime(2009, 6, 30);
       Console.WriteLine($"D Format Specifier:     {date1:D}");
       string longPattern = CultureInfo.CurrentCulture.DateTimeFormat.LongDatePattern;
-      Console.WriteLine("'{0}' custom format string:     {1}",
-                        longPattern, date1.ToString(longPattern));
+      Console.WriteLine($"'{longPattern}' custom format string:     {date1.ToString(longPattern)}");
    }
 }
 // The example displays the following output when run on a system whose
