@@ -1,4 +1,4 @@
-﻿// <Snippet18>
+// <Snippet18>
 using System;
 using System.Globalization;
 using System.Threading;
@@ -9,8 +9,8 @@ public class Example17
     {
         Thread.CurrentThread.CurrentCulture = CultureInfo.CreateSpecificCulture("pl-PL");
         string composite = "\u0041\u0300";
-        Console.WriteLine("Comparing using Char:   {0}", composite.IndexOf('\u00C0'));
-        Console.WriteLine("Comparing using String: {0}", composite.IndexOf("\u00C0"));
+        Console.WriteLine($"Comparing using Char:   {composite.IndexOf('\u00C0')}");
+        Console.WriteLine($"Comparing using String: {composite.IndexOf("\u00C0")}");
     }
 }
 
