@@ -1,4 +1,4 @@
-﻿// <Snippet3>
+// <Snippet3>
 using System;
 using System.Text.RegularExpressions;
 
@@ -9,8 +9,7 @@ public class Class1
       string pattern = @"\b(\w+?)\s\1\b";
       string input = "This this is a nice day. What about this? This tastes good. I saw a a dog.";
       foreach (Match match in Regex.Matches(input, pattern, RegexOptions.IgnoreCase))
-         Console.WriteLine("{0} (duplicates '{1}') at position {2}",
-                           match.Value, match.Groups[1].Value, match.Index);
+         Console.WriteLine($"{match.Value} (duplicates '{match.Groups[1].Value}') at position {match.Index}");
    }
 }
 // The example displays the following output:
