@@ -12,7 +12,7 @@ class SecurityPrincipalDemo
 
         myDomain.SetPrincipalPolicy(PrincipalPolicy.WindowsPrincipal);
         WindowsPrincipal myPrincipal = (WindowsPrincipal)Thread.CurrentPrincipal;
-        Console.WriteLine("{0} belongs to: ", myPrincipal.Identity.Name.ToString());
+        Console.WriteLine($"{myPrincipal.Identity.Name.ToString()} belongs to: ");
         Array wbirFields = Enum.GetValues(typeof(WindowsBuiltInRole));
         //<Snippet2>
         foreach (object roleName in wbirFields)
@@ -20,8 +20,7 @@ class SecurityPrincipalDemo
             try
             {
                 // Cast the role name to a RID represented by the WindowsBuildInRole value.
-                Console.WriteLine("{0}? {1}.", roleName,
-                    myPrincipal.IsInRole((WindowsBuiltInRole)roleName));
+                Console.WriteLine($"{roleName}? {myPrincipal.IsInRole((WindowsBuiltInRole)roleName)}.");
                 Console.WriteLine("The RID for this role is: " + ((int)roleName).ToString());
             }
             catch (Exception)
@@ -32,20 +31,17 @@ class SecurityPrincipalDemo
         //</Snippet2>
         //<Snippet3>
         // Get the role using the string value of the role.
-        Console.WriteLine("{0}? {1}.", "Administrators",
-            myPrincipal.IsInRole("BUILTIN\\" + "Administrators"));
-        Console.WriteLine("{0}? {1}.", "Users",
-            myPrincipal.IsInRole("BUILTIN\\" + "Users"));
+        Console.WriteLine($"'Administrators'? {myPrincipal.IsInRole("BUILTIN\\" + "Administrators")}.");
+        Console.WriteLine($"'Users'? {myPrincipal.IsInRole("BUILTIN\\" + "Users")}.");
         //</Snippet3>
         //<Snippet4>
         // Get the role using the WindowsBuiltInRole enumeration value.
-        Console.WriteLine("{0}? {1}.", WindowsBuiltInRole.Administrator,
-           myPrincipal.IsInRole(WindowsBuiltInRole.Administrator));
+        Console.WriteLine($"{WindowsBuiltInRole.Administrator}? {myPrincipal.IsInRole(WindowsBuiltInRole.Administrator)}.");
         //</Snippet4>
         //<Snippet5>
         // Get the role using the WellKnownSidType.
         SecurityIdentifier sid = new SecurityIdentifier(WellKnownSidType.BuiltinAdministratorsSid, null);
-        Console.WriteLine("WellKnownSidType BuiltinAdministratorsSid  {0}? {1}.", sid.Value, myPrincipal.IsInRole(sid));
+        Console.WriteLine($"WellKnownSidType BuiltinAdministratorsSid  {sid.Value}? {myPrincipal.IsInRole(sid)}.");
         //</snippet5>
     }
 
