@@ -1,4 +1,4 @@
-﻿// This older code snippet replaced by a new one
+// This older code snippet replaced by a new one
 // This code example demonstrates the ToString(String) and
 // ToString(String, IFormatProvider) methods for integral and
 // floating-point numbers, in conjunction with the standard
@@ -182,40 +182,25 @@ public class NumericFormats
 
       // Output floating point values
       double floating = 10761.937554;
-      Console.WriteLine("C: {0}",
-              floating.ToString("C", ci));           // Displays "C: $10,761.94"
-      Console.WriteLine("E: {0}",
-              floating.ToString("E03", ci));         // Displays "E: 1.076E+004"
-      Console.WriteLine("F: {0}",
-              floating.ToString("F04", ci));         // Displays "F: 10761.9376"
-      Console.WriteLine("G: {0}",
-              floating.ToString("G", ci));           // Displays "G: 10761.937554"
-      Console.WriteLine("N: {0}",
-              floating.ToString("N03", ci));         // Displays "N: 10,761.938"
-      Console.WriteLine("P: {0}",
-              (floating/10000).ToString("P02", ci)); // Displays "P: 107.62 %"
-      Console.WriteLine("R: {0}",
-              floating.ToString("R", ci));           // Displays "R: 10761.937554"
+      Console.WriteLine($"C: {floating.ToString("C", ci)}");           // Displays "C: $10,761.94"
+      Console.WriteLine($"E: {floating.ToString("E03", ci)}");         // Displays "E: 1.076E+004"
+      Console.WriteLine($"F: {floating.ToString("F04", ci)}");         // Displays "F: 10761.9376"
+      Console.WriteLine($"G: {floating.ToString("G", ci)}");           // Displays "G: 10761.937554"
+      Console.WriteLine($"N: {floating.ToString("N03", ci)}");         // Displays "N: 10,761.938"
+      Console.WriteLine($"P: {(floating/10000).ToString("P02", ci)}"); // Displays "P: 107.62 %"
+      Console.WriteLine($"R: {floating.ToString("R", ci)}");           // Displays "R: 10761.937554"
       Console.WriteLine();
 
       // Output integral values
       int integral = 8395;
-      Console.WriteLine("C: {0}",
-              integral.ToString("C", ci));           // Displays "C: $8,395.00"
-      Console.WriteLine("D: {0}",
-              integral.ToString("D6", ci));          // Displays "D: 008395"
-      Console.WriteLine("E: {0}",
-              integral.ToString("E03", ci));         // Displays "E: 8.395E+003"
-      Console.WriteLine("F: {0}",
-              integral.ToString("F01", ci));         // Displays "F: 8395.0"
-      Console.WriteLine("G: {0}",
-              integral.ToString("G", ci));           // Displays "G: 8395"
-      Console.WriteLine("N: {0}",
-              integral.ToString("N01", ci));         // Displays "N: 8,395.0"
-      Console.WriteLine("P: {0}",
-              (integral/10000.0).ToString("P02", ci)); // Displays "P: 83.95 %"
-      Console.WriteLine("X: 0x{0}",
-              integral.ToString("X", ci));           // Displays "X: 0x20CB"
+      Console.WriteLine($"C: {integral.ToString("C", ci)}");           // Displays "C: $8,395.00"
+      Console.WriteLine($"D: {integral.ToString("D6", ci)}");          // Displays "D: 008395"
+      Console.WriteLine($"E: {integral.ToString("E03", ci)}");         // Displays "E: 8.395E+003"
+      Console.WriteLine($"F: {integral.ToString("F01", ci)}");         // Displays "F: 8395.0"
+      Console.WriteLine($"G: {integral.ToString("G", ci)}");           // Displays "G: 8395"
+      Console.WriteLine($"N: {integral.ToString("N01", ci)}");         // Displays "N: 8,395.0"
+      Console.WriteLine($"P: {(integral/10000.0).ToString("P02", ci)}"); // Displays "P: 83.95 %"
+      Console.WriteLine($"X: 0x{integral.ToString("X", ci)}");           // Displays "X: 0x20CB"
       Console.WriteLine();
       // </SnippetFinalExample>
    }
